fix: snowGolemAttackRechargeSpeedMultiplier cast missing

+++ src/main/java/net/scirave/nox/mixin/SnowGolemEntityMixin.java
@@ -23,7 +23,7 @@ public abstract class SnowGolemEntityMixin extends GolemEntityMixin {
 
     @ModifyArgs(method = "initGoals", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/ai/goal/ProjectileAttackGoal;<init>(Lnet/minecraft/entity/ai/RangedAttackMob;DIF)V"))
     public void nox$snowGolemFasterShooting(Args args) {
-        args.set(2, ((int) args.get(2)) / Math.max(NoxConfig.snowGolemAttackRechargeSpeedMultiplier, 0));
+        args.set(2, (int)(((int) args.get(2)) / Math.max(NoxConfig.snowGolemAttackRechargeSpeedMultiplier, 0)));
         args.set(3, ((float) args.get(3)) * Math.max(NoxConfig.snowGolemAttackRangeMultiplier, 1));
     }
     @ModifyArgs(method = "attack", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/projectile/thrown/SnowballEntity;setVelocity(DDDFF)V"))
